Shuffle directory structure a bit.
Moved the tests to the ChessTools/Test/ directory after thinking about
how the test suite is likely to develop. This should help avoid the need
for clunky naming to avoid module name clashes when importing (it won't
make the full import name any shorter, but it will be more descriptive
and obviously hierarchicaly).

Thus,
	TestBoard -> ChessTools.Test.Board, and
	Suite -> ChessTools.Test.Suite.

Necessary modifications made to the source so that the tests are still
all passing.